# Optimization: Cache WorldEdit schematic