Test on Windows 11 x64 : 
Java 16.0.1
Purpur 1.17.1 build 1400
WorldEdit 7.2.6 (No FAWE, No AWE)
I9-9900K (4 vcpu)

Without cache : 30 ms and 17 tps and 4 sec for teleport player
With cache :
- first paste : 30 ms / 17 tps / 4 sec for teleport player
- after : 10 ms / 20 tps and ~1 sec for teleport player